% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{load_all_courses}
\alias{load_all_courses}
\title{load all Swirl courses}
\usage{
load_all_courses(directory)
}
\arguments{
\item{directory}{path to directory where all the subdirectories are course directories}
}
\description{
load all Swirl courses. Courses can be
created as described here: http://swirlstats.com/instructors.html.
}
\details{
It returns a combined data frame for all courses
}
