% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{load_course}
\alias{load_course}
\title{load a Swirl course}
\usage{
load_course(directory)
}
\arguments{
\item{directory}{path to directory of course to upload}
}
\description{
load a Swirl course. Courses can be
created as described here: http://swirlstats.com/instructors.html.
}
\details{
It returns the a data frame.
}
