% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{load_student_list}
\alias{load_student_list}
\title{load all Swirl courses}
\usage{
load_student_list(file)
}
\arguments{
\item{file}{that stores the merged records}
}
\description{
load all Swirl courses
}
\details{
It returns a combined data frame for the merged
}
